Fixed some services and dao classes

+++ server-model/src/main/java/epam/training/finalproject/model/service/impls/OrderServiceImpl.java
@@ -1,12 +1,13 @@
 package epam.training.finalproject.model.service.impls;
 
 import epam.training.finalproject.exceptions.EntityNotFoundException;
+import epam.training.finalproject.exceptions.OperationException;
 import epam.training.finalproject.model.dao.interfaces.OrderDao;
 import epam.training.finalproject.model.domain.entity.Order;
 import epam.training.finalproject.model.domain.entity.enums.OrderStatus;
 import epam.training.finalproject.model.service.interfaces.OrderService;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,72 +15,91 @@ import java.util.List;
 @Service
 public class OrderServiceImpl implements OrderService {
 
-    private final long orderDurationTimeOut=30;
+    private static final Logger LOGGER = Logger.getLogger(OrderServiceImpl.class);
+
+    private final long orderDurationTimeOut = 30;
 
     @Autowired
     private OrderDao orderDao;
 
-    public List<Order> findOrdersByUserId(long idUser){
-        try {
-            return orderDao.findOrdersByUserId(idUser);
+    public List<Order> findOrdersByUserId(Long userId) {
+        if (userId > 0) {
+            List<Order> orders = orderDao.findOrdersByUserId(userId);
+            if (orders.isEmpty()){
+                LOGGER.debug("Orders is empty!");
+                throw new EntityNotFoundException("Any orders are not found for user with id " + userId);
+            }
+            return orders;
         }
-        catch (DataAccessException ex){
-            throw new EntityNotFoundException("Order for user with id "+idUser+" is not found",ex.getCause());
-        }
+        LOGGER.debug("User id is invalid!");
+        throw new IllegalArgumentException("User id is invalid");
     }
-//check this
+
     @Override
     public Order getById(Long id) {
-        try {
-            return orderDao.getById(id).get();
+        if (id > 0) {
+            return orderDao.getById(id).orElseThrow(() -> {
+                LOGGER.debug("Order is null!");
+                return new EntityNotFoundException("Order with id " + id + " is not found");
+            });
         }
-        catch (DataAccessException ex){
-            throw new EntityNotFoundException("Order with id "+id+" is not found",ex.getCause());
-        }
+        LOGGER.debug("Order id is invalid!");
+        throw new IllegalArgumentException("Order id is invalid");
     }
 
     @Override
     public List<Order> getAll() {
-        try {
-            return orderDao.getAll();
+        List<Order> orders = orderDao.getAll();
+        if (orders.isEmpty()){
+            LOGGER.debug("Orders is empty!");
+            throw new EntityNotFoundException("Any orders are not found");
         }
-        catch (DataAccessException ex){
-            throw new EntityNotFoundException("Orders are not found",ex.getCause());
-        }
+        return orders;
     }
 
+    //TODO add logic to delete/update operations (example save())
     @Override
     public Long delete(Order order) {
-        if (order!=null)
+        if (order != null && order.getStatus().equals(OrderStatus.FINISHED)) {
+            order.setDeleted(true);
             return orderDao.delete(order);
-        return 0L;
+        }
+        throw new IllegalArgumentException("Order is invalid");
     }
 
     @Override
     public Long updateOrderSetPaid(Order order) {
-        if (order!=null && order.getPaymentDate().isBefore(order.getConfirmationDate().plusMinutes(orderDurationTimeOut)) && order.getStatus()== OrderStatus.PAID)
+        if (order != null && order.getPaymentDate().isBefore(order.getConfirmationDate().plusMinutes(orderDurationTimeOut)) && order.getStatus() == OrderStatus.PAID)
             return orderDao.updateOrderSetPaid(order);
-        return 0L;
+        throw new IllegalArgumentException("Order is invalid");
     }
 
     @Override
     public Long updateOrderStatus(Order order) {
-        if (order!=null)
+        if (order != null)
             return orderDao.updateOrderStatus(order);
-        return 0L;
+        throw new IllegalArgumentException("Order is invalid");
     }
 
     @Override
     public Long update(Order order) {
-        if (order!=null && order.getConfirmationDate()!=null && order.getStatus()==OrderStatus.CONFIRMED && order.getPrice()>0)
+        if (order != null && order.getConfirmationDate() != null && order.getStatus() == OrderStatus.CONFIRMED && order.getPrice() > 0)
             return orderDao.update(order);
-        return 0L;
+        throw new IllegalArgumentException("Order is invalid");
     }
 
     @Override
     public Long save(Order order) {
-        if (order!=null && order.getOrderDuration()>0 && order.getStatus()== OrderStatus.CREATED && order.getUserId()>0 &&  order.getCarId()>0)
-            return orderDao.save(order);
-        return 0L;
+        if (order != null){
+            order.setStatus(OrderStatus.CREATED);
+            Long result = orderDao.save(order);
+            if (result == -1L){
+                LOGGER.debug("Invalid order credentials!");
+                throw new OperationException("Order with id " + order.getId() + " has invalid credentials");
+            }
+            return result;
+        }
+        LOGGER.debug("Order is null!");
+        throw new IllegalArgumentException("Order is invalid");
     }
 }
